requirements.txt
streamlit
Pillow
tensorflow
numpy
matplotlib
scikit-learn
